feat(jwt): acceptable CryptoKey in JWT sign and verify (#2373)
* supported jwt verification with private key

* created jwa.ts and moved some functions

* added jws.ts and removed key.ts

* acceptable CryptoKey object

* denoify

* fixed jwt middleware tests

* fixed handling CryptoKey for node v18
