+++ deno_dist/utils/jwt/index.ts
@@ -1,1 +1,2 @@
-export * as Jwt from './jwt.ts'
+import { sign, verify, decode } from './jwt.ts'
+export const Jwt = { sign, verify, decode }
