+++ src/utils/jwt/jwt.ts
@@ -1,14 +1,14 @@
 import { encodeBase64Url, decodeBase64Url } from '../../utils/encode'
-import { importPrivateKey, importPublicKey } from './key'
-import type { AlgorithmParams, TokenHeader, AlgorithmTypeName } from './types'
-import { JwtHeaderInvalid, JwtTokenIssuedAt, isTokenHeader } from './types'
+import type { SignatureAlgorithm } from './jwa'
+import { AlgorithmTypes } from './jwa'
+import type { SignatureKey } from './jws'
+import { signing, verifying } from './jws'
+import { JwtHeaderInvalid, JwtTokenIssuedAt } from './types'
 import {
   JwtTokenInvalid,
   JwtTokenNotBefore,
   JwtTokenExpired,
   JwtTokenSignatureMismatched,
-  JwtAlgorithmNotImplemented,
-  CryptoKeyUsage,
 } from './types'
 import { utf8Decoder, utf8Encoder } from './utf8'
 
@@ -19,151 +19,43 @@ const encodeSignaturePart = (buf: ArrayBufferLike): string => encodeBase64Url(bu
 const decodeJwtPart = (part: string): unknown =>
   JSON.parse(utf8Decoder.decode(decodeBase64Url(part)))
 
-const param = (name: AlgorithmTypeName): AlgorithmParams => {
-  switch (name) {
-    case 'HS256':
-      return {
-        name: 'HMAC',
-        hash: {
-          name: 'SHA-256',
-        },
-      }
-    case 'HS384':
-      return {
-        name: 'HMAC',
-        hash: {
-          name: 'SHA-384',
-        },
-      }
-    case 'HS512':
-      return {
-        name: 'HMAC',
-        hash: {
-          name: 'SHA-512',
-        },
-      }
-    case 'RS256':
-      return {
-        name: 'RSASSA-PKCS1-v1_5',
-        hash: {
-          name: 'SHA-256',
-        },
-      }
-    case 'RS384':
-      return {
-        name: 'RSASSA-PKCS1-v1_5',
-        hash: {
-          name: 'SHA-384',
-        },
-      }
-    case 'RS512':
-      return {
-        name: 'RSASSA-PKCS1-v1_5',
-        hash: {
-          name: 'SHA-512',
-        },
-      }
-    case 'PS256':
-      return {
-        name: 'RSA-PSS',
-        hash: {
-          name: 'SHA-256',
-        },
-        saltLength: 32, // 256 >> 3
-      } satisfies RsaPssParams & RsaHashedImportParams
-    case 'PS384':
-      return {
-        name: 'RSA-PSS',
-        hash: {
-          name: 'SHA-384',
-        },
-        saltLength: 48, // 384 >> 3
-      } satisfies RsaPssParams & RsaHashedImportParams
-    case 'PS512':
-      return {
-        name: 'RSA-PSS',
-        hash: {
-          name: 'SHA-512',
-        },
-        saltLength: 64, // 512 >> 3,
-      } satisfies RsaPssParams & RsaHashedImportParams
-    case 'ES256':
-      return {
-        name: 'ECDSA',
-        hash: {
-          name: 'SHA-256',
-        },
-        namedCurve: 'P-256',
-      } satisfies EcdsaParams & EcKeyImportParams
-    case 'ES384':
-      return {
-        name: 'ECDSA',
-        hash: {
-          name: 'SHA-384',
-        },
-        namedCurve: 'P-384',
-      } satisfies EcdsaParams & EcKeyImportParams
-    case 'ES512':
-      return {
-        name: 'ECDSA',
-        hash: {
-          name: 'SHA-512',
-        },
-        namedCurve: 'P-521',
-      } satisfies EcdsaParams & EcKeyImportParams
-    case 'EdDSA':
-      // Currently, supported only Safari and Deno, Node.js.
-      // See: https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/verify
-      return {
-        name: 'Ed25519',
-        namedCurve: 'Ed25519',
-      }
-    default:
-      throw new JwtAlgorithmNotImplemented(name)
-  }
+export interface TokenHeader {
+  alg: SignatureAlgorithm
+  typ: 'JWT'
 }
 
-const signing = async (
-  data: string,
-  privateKey: string | JsonWebKey,
-  alg: AlgorithmTypeName = 'HS256'
-): Promise<ArrayBuffer> => {
-  const algorithm = param(alg)
-  const cryptoKey = await importPrivateKey(privateKey, algorithm, [CryptoKeyUsage.Sign])
-  return await crypto.subtle.sign(algorithm, cryptoKey, utf8Encoder.encode(data))
+// eslint-disable-next-line
+export function isTokenHeader(obj: any): obj is TokenHeader {
+  return (
+    typeof obj === 'object' &&
+    obj !== null &&
+    'alg' in obj &&
+    Object.values(AlgorithmTypes).includes(obj.alg) &&
+    'typ' in obj &&
+    obj.typ === 'JWT'
+  )
 }
 
 export const sign = async (
   payload: unknown,
-  privateKey: string | JsonWebKey,
-  alg: AlgorithmTypeName = 'HS256'
+  privateKey: SignatureKey,
+  alg: SignatureAlgorithm = 'HS256'
 ): Promise<string> => {
   const encodedPayload = encodeJwtPart(payload)
   const encodedHeader = encodeJwtPart({ alg, typ: 'JWT' } satisfies TokenHeader)
 
   const partialToken = `${encodedHeader}.${encodedPayload}`
 
-  const signaturePart = await signing(partialToken, privateKey, alg)
+  const signaturePart = await signing(privateKey, alg, utf8Encoder.encode(partialToken))
   const signature = encodeSignaturePart(signaturePart)
 
   return `${partialToken}.${signature}`
 }
 
-const verifying = async (
-  publicKey: string | JsonWebKey,
-  alg: AlgorithmTypeName = 'HS256',
-  signature: BufferSource,
-  data: BufferSource
-): Promise<boolean> => {
-  const algorithm = param(alg)
-  const cryptoKey = await importPublicKey(publicKey, algorithm, [CryptoKeyUsage.Verify])
-  return await crypto.subtle.verify(algorithm, cryptoKey, signature, data)
-}
-
 export const verify = async (
   token: string,
-  publicKey: string | JsonWebKey,
-  alg: AlgorithmTypeName = 'HS256'
+  publicKey: SignatureKey,
+  alg: SignatureAlgorithm = 'HS256'
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
 ): Promise<any> => {
   const tokenParts = token.split('.')
